feat: added error handling for all pages and added get router for share link

+++ controllers/deleteFile.js
@@ -6,26 +6,36 @@ const { promisify } = require("util");
 const unlinkAsync = promisify(fs.unlink);
 
 async function deleteFile(req, res, next) {
-  const { name, location } = req.query;
-  const file = await prisma.files.findFirst({
-    where: {
-      name: name,
-      location: location,
-      usersId: req.user.id,
-    },
-  });
+  const auth = req.isAuthenticated();
 
-  await unlinkAsync(file.path);
-  await prisma.files.delete({
-    where: {
-      id: file.id,
-    },
-  });
+  if (auth) {
+    try {
+      const { name, location } = req.query;
+      const file = await prisma.files.findFirst({
+        where: {
+          name: name,
+          location: location,
+          usersId: req.user.id,
+        },
+      });
 
-  if (location === `/`) {
-    res.redirect(`/vault`);
+      await unlinkAsync(file.path);
+      await prisma.files.delete({
+        where: {
+          id: file.id,
+        },
+      });
+
+      if (location === `/`) {
+        res.redirect(`/vault`);
+      } else {
+        res.redirect(`/vault/${location}`);
+      }
+    } catch (err) {
+      console.log(err);
+    }
   } else {
-    res.redirect(`/vault/${location}`);
+    res.redirect(`/login`);
   }
 }
